backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
crewai
langchain
chromadb
python-dotenv
httpx
pytest
pyinstaller
langgraph
langchain
faiss-cpu
fastapi
uvicorn
httpx
openai
langchain
python-dotenv
tiktoken
faiss-cpu
jinja2
